[test] quest & dialogue
test connecting quest & dialogue on MJ Scene
