--- Igloo/Assets/Code/Scripts/Dialogue.cs
@@ -1,11 +1,0 @@
-using System.Collections;
-using System.Collections.Generic;
-using UnityEngine;
-
-[System.Serializable]
-public class Dialogue
-{
-    public string name;
-    [TextArea(3, 10)]
-    public string[] sentences;
-}
