--- Igloo/Assets/Code/Scripts/Dialogue.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 3c8b55cf2f62aa74d848929eed41b15d
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
