+++ Igloo/Assets/Code/Scripts/Dialogue.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 0ceb1f915c5b0774091234d1bf86b536
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
