+++ Igloo/Assets/Code/Scripts/Dialogue/Dialogue.cs
@@ -1,0 +1,11 @@
+using System.Collections;
+using System.Collections.Generic;
+using UnityEngine;
+
+[System.Serializable]
+public class Dialogue
+{
+    public string name;
+    [TextArea(3, 10)]
+    public string[] sentences;
+}
