--- Igloo/Assets/Code/Scripts/DialogueManager.cs
@@ -1,62 +1,0 @@
-using System.Collections;
-using System.Collections.Generic;
-using UnityEngine;
-using UnityEngine.UI;
-using TMPro;
-
-public class DialogueManager : MonoBehaviour
-{
-    public TMP_Text nameText;
-    public TMP_Text dialogueText;
-    public Animator animator;
-    private Queue<string> sentences;
-
-    void Start()
-    {
-        sentences = new Queue<string>();
-    }
-
-    void Update()
-    {
-
-    }
-
-    public void StartDialogue(Dialogue dialogue)
-    {
-        animator.SetBool("IsOpen", true);
-        nameText.text = dialogue.name;
-        sentences.Clear();
-        foreach (string sentence in dialogue.sentences)
-        {
-            sentences.Enqueue(sentence);
-        }
-        DisplayNextSentence();
-    }
-
-    public void DisplayNextSentence()
-    {
-        if (sentences.Count == 0)
-        {
-            EndDialogue();
-            return;
-        }
-        string sentence = sentences.Dequeue();
-        StopAllCoroutines();
-        StartCoroutine(TypeSentence(sentence));
-    }
-
-    void EndDialogue()
-    {
-        animator.SetBool("IsOpen", false);
-    }
-
-    IEnumerator TypeSentence(string sentence)
-    {
-        dialogueText.text = "";
-        foreach (char letter in sentence.ToCharArray())
-        {
-            dialogueText.text += letter;
-            yield return null;
-        }
-    }
-}
