--- Igloo/Assets/Code/Scripts/DialogueManager.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: dc33156035040884583821da7cb377fc
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
