--- Igloo/Assets/Code/Scripts/DialogueTrigger.cs
@@ -1,65 +1,0 @@
-using System.Collections;
-using System.Collections.Generic;
-using UnityEngine;
-
-public class DialogueTrigger : MonoBehaviour
-{
-    public Dialogue dialogue;
-    public GameObject visualCue;
-    private bool isPlayerInRange;
-
-    private void Awake()
-    {
-        isPlayerInRange = false;
-        visualCue.SetActive(false);
-    }
-
-    private void Update()
-    {
-        if (isPlayerInRange)
-        {
-            visualCue.SetActive(true);
-        }
-        else
-        {
-            visualCue.SetActive(false);
-        }
-    }
-
-    private void OnTriggerEnter(Collider collider)
-    {
-        if (collider.gameObject.tag == "Player")
-        {
-            isPlayerInRange = true;
-        }
-    }
-
-    private void OnTriggerStay(Collider collider)
-    {
-        if  (collider.gameObject.tag == "Player")
-        {
-            if (Input.GetKeyDown(KeyCode.E))
-            {
-                TriggerDialogue();
-            }
-            if (Input.GetKeyDown(KeyCode.Space))
-            {
-                FindObjectOfType<DialogueManager>().DisplayNextSentence();
-            }
-        }
-    }
-
-    private void OnTriggerExit(Collider collider)
-    {
-        if (collider.gameObject.tag == "Player")
-        {
-            isPlayerInRange = false;
-        }
-    }
-
-    public void TriggerDialogue()
-    {
-        FindObjectOfType<DialogueManager>().StartDialogue(dialogue);
-    }
-
-}
