--- Igloo/Assets/Code/Scripts/DialogueTrigger.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 3828f80f743fe144bbfa1872f013130d
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
