+++ Igloo/Assets/Resources/Quests/TutorialQuest.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 2f9c3603620b57a42b5f725ff4c1f6a1
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
