+++ Igloo/Assets/Resources/Quests/TutorialQuest/GoToUnivStep.cs
@@ -1,0 +1,14 @@
+using System.Collections;
+using System.Collections.Generic;
+using UnityEngine;
+
+public class GoToUnivStep : QuestStep
+{
+    private void OnTriggerEnter(Collider other)
+    {
+        if (other.gameObject.CompareTag("Player"))
+        {
+            FinishQuestStep();
+        }
+    }
+}
