+++ Igloo/Assets/Resources/Quests/TutorialQuest/GoToUnivStep.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 581241cf2619aa244a550a8432a0d739
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
