+++ Igloo/Assets/Resources/Quests/TutorialQuest/TutorialQuest.asset
@@ -1,0 +1,20 @@
+%YAML 1.1
+%TAG !u! tag:unity3d.com,2011:
+--- !u!114 &11400000
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 0}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: 3c4507349286d484caed246f9f5bae50, type: 3}
+  m_Name: TutorialQuest
+  m_EditorClassIdentifier: 
+  <id>k__BackingField: TutorialQuest
+  displayName: Tutorial
+  description: "\uD29C\uD1A0\uB9AC\uC5BC \uD018\uC2A4\uD2B8\uB97C \uC9C4\uD589\uD574\uBCF4\uC790."
+  questPrerequisites: []
+  questStepPrefabs: []
+  reward: {fileID: 0}
