+++ Igloo/Assets/Resources/Quests/TutorialQuest/TutorialQuest.asset.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 6db6aa7b62884a4448fb75d20fb043ca
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 11400000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
